# [NES][FFI] Optimization