--- a/apps/NES/_CoqProject
+++ b/apps/NES/_CoqProject
@@ -9,5 +9,7 @@
 theories/NES.v
 
 tests/test_NES.v
+tests/test_NES_perf.v
+tests/test_NES_perf_optimal.v
 
 examples/usage_NES.v