--- a/src/coq_elpi_builtins.ml
+++ b/src/coq_elpi_builtins.ml
@@ -413,7 +413,7 @@ The name and the grafting specification can be left unspecified.|};
 } |> CConv.(!<)
 
 let set_accumulate_to_db, get_accumulate_to_db =
-  let f = ref ((fun () -> assert false),(fun _ _ ~local:_ -> assert false)) in
+  let f = ref ((fun () -> assert false),(fun _ _ ~local:_ -> assert false),(fun () -> assert false)) in
   (fun x -> f := x),
   (fun () -> !f)
 
@@ -2050,8 +2050,8 @@ Supported attributes:
          State.update clauses_for_later state (fun l ->
            (dbname,clause,local) :: l), (), []
      | Given CurrentModule ->
-          let elpi, f = get_accumulate_to_db () in
-          f dbname API.(Compile.unit ~elpi:(elpi ()) ~flags:Compile.default_flags clause) ~local;
+          let elpi, f, cur_program = get_accumulate_to_db () in
+          f dbname API.(Compile.unit ~follows:(cur_program ()) ~elpi:(elpi ()) ~flags:Compile.default_flags clause) ~local;
           state, (), []
      )),
   DocAbove);