The test_NES_perf file runs Elpi 1000 times. It takes 32 seconds vs 0.5 seconds running "the same" Coq code. The speed of Elpi does not seem to be the culprit. Some API are way slower than their ML counterpart (this overhead is repeated 2K times for begin-module, 2K times for end-module).

Investigation:
- `UState.demote_global_univs` takes 0.0007s on top of APIs taking 0.00007s (around 3s total I guess)
- Elpi's `compiler` (23s in total)
  - on the initial program
    - assembling: 0.0089
    - optimization:0.0062
  - on the last program (500+ ns clauses/compilation units)
    - assembling:0.0406
    - optimization:0.0033
- Elpi's `runtime` (7s in total, including the 3s API overhead)