@gares 2020-09-01
By using the new API for program extension in https://github.com/LPCIC/elpi/pull/85 we can implement a cache for programs and quickly append to them clauses (DB entries).

We are now at 11s.
If one disables the `UState.demote_global_univs` call, we are at 6s, so I guess the focus should be there now.

I believe 10x slower than Coq is OK, in the sense that is the speed factor between a compiled and interpreted language.